python: vendor parsedatetime
Problem: Parsing user input that specifies date and time or a time
offset is an important but onerous task. It is simple to support
restricted datetime formats (e.g. ISO 8061), but it is much more
user-friendly to allow a wide range of input including partial dates
with missing information like "Jun 7" and even natural language inputs
like "next Thursday" or "an hour ago".

It turns out there are several existing Python modules to do this
and 'parsedatetime' appears to be one of the best. However, we
don't want to add a new external dependency to flux-core for this
limited purpose, so vendor the latest version of parsedatetime under
'flux.utils.parsedatetime'.

https://github.com/bear/parsedatetime
:
 commit c55337589ee582813182b74f2d3ae80e2fcd9738
 Author: bear (Mike Taylor) <bear@bear.im>
 Date:   Sun May 31 19:24:18 2020 -0400