python: catch FluxExecutor thread exceptions
When a FluxExecutor thread crashes, the executor becomes
unusable and Python hangs. Threads can crash in certain
specific circumstances---for instance, when the executor
is connected to a remote Flux instance which has gone away.

Make it so that when a thread raises an exception, all
outstanding futures have that exception set on them.