github: fix the fix for ci builds on master
Problem: The fix for builds on master didn't work, since GITHUB_BRANCH
was not set in the environment of the annotated tags step (it was set
as an output).

Fall back to using github.ref, and skip the annotated tags checkout
when github.ref == 'refs/heads/master'.

Add a line of debugging in case this step starts to fail again in the
future.