libflux: make topic member of flux_plugin_handler const
Problem: The topic string in `struct flux_plugin_handler` is not
const, which causes compilation errors when using static initializers
to define a table of handlers on compilers where static strings
are const (e.g. C++11).

Make the topic string const. The only reason it wasn't const was as a
convenience when destroying copied flux_plugin_handler structs. Just
cast the topic back to non-const when freeing the copy.

Fixes #3719