testsuite: disable validator when submitting invalid jobspec
Problem: A recent update to the Jobspec class rejects jobspec with
invalid dependencies, but a couple tests purposefuly do this to
check that dependencies plugins do not fail when presented with
invalid input.

Disable the validator during these tests to ensure the invalid
input makes it to the plugin being tested.