cron: do not start task timeout until task is running
Problem: Timeouts for cron tasks are started as soon as flux_rexec()
returns, but if the timer fires before the task is running (i.e.
the flux_rexec() future is fulfilled), then the task cannot be
killed in the timer callback. This can cause the tasks to remain
running when they should have been terminated.

Do not start the timeout timer watcher until the task is in the
running state, so there is no possibility of an early timeout
trying to kill a task that is still starting.