testsuite: jobspec/validate.py: give more specific errors
Problem: When there is a JSON decode error in the jsonschema input
to jobspec/validate.py, the script gives a generic "unknown error",
which is not helpful.

Catch JSONDecodeError specifically and call out the error type, otherwise
print the string representation of the Exception in some hopes it will
be useful to the end user.