t: t2608-job-shell-log.t: fix false failures
Problem: In t2608-job-shell-log.t tests that fatal errors terminate
jobs only allow the jobs to either fail or be killed by SIGTERM,
but on slow systems the jobs may appear as killed by SIGKILL since
this signal is used as a last resort. This causes false failures
in this test on some systems.

Enhance test_must_fail_or_be_teriminated() to additionally allow
jobs to be killed by SIGKILL to workaround this issue.