sharness: add test_must_fail_or_be_terminated()
Problem: In some cases we want to use sharness' test_must_fail(),
but also allow the program to be terminated by SIGTERM or SIGKILL.
In t2608-job-shell-log.t there is an existing shell function that
does what we want, but it would be good to avoid copy-and-paste of
this shell function.

Move test_must_fail_or_be_terminated() from t2608-job-shell-log.t to
sharness.d/flux-sharness.sh so that it can be used from all tests.