t: t2900-job-timelimits.t: workaround racy timeout test
Problem: The "sigalrm_sigkill_test" intermittently fails because the
underlying job exits with a returncode of 1 instead of being killed
by SIGKILL. This could be due to when the signal is delivered to the
underlying job script. In any case the key here is that SIGALRM is
received before the job terminates, so these are false test failures.

Use test_must_fail_or_be_terminated() instead of test_expect_code()
in the test to workaround this failure case.

Also adjust the test so that separate runs of the test use different
output and error files for enhanced debugging.