flux-start: add --test-rundir-cleanup
Problem: rundir is not cleaned up if it was specified on the
command line, as opposed to created internally, however the
sharness "system" personality needs to specify the rundir in
advance to use in configured ipc paths, but is simplified by
letting flux-start clean up as it does with other test personalities.

Add a --test-rundir-cleanup option to allow cleanup to be requested.

Update flux-start(1).
Test option in t0001-basic.t.