testsuite: use short ipc:// paths
Problem: the system test personality places ipc:// sockets
in the test trash directory, but if the flux tree path is
long, this can overflow sockaddr_un.sun_path[], which on
linux is around 100 bytes.

The ipc paths need to be specified in the config file that
is created before flux-start runs.  Create a temporary broker
"rundir" in /tmp (or $TMPDIR) for the test and place the sockets
there, then run flux start with --test-rundir=DIR --test-rundir-cleanup.

Fixes #3736