job-manager: add alloc_bypass job flag
Problem: In some cases, the job manager may want to bypass the
scheduler alloc protocol and assign an R manually to a job, but
there is no current way to accomplish this.

Add a new job->alloc_bypass flag which disables the alloc request to
the scheduler. The job still must get an "alloc" event before
proceeding from the RUN to the SCHED state, and a "free" event before
it can proceed from CLEANUP to INACTIVE. The entity setting the
alloc_bypass flag will be responsible for ensuring these events
are emitted at the appropriate time.