job-manager: add alloc-bypass.so plugin
Problem: While the job-manager supports bypass of the scheduler via a
the 'alloc_bypass' flag, there is no current method to take advantage
of this capability.

Add an alloc-bypass.so plugin which looks for

  attributes.system.alloc-bypass.R

set in jobspec. If this object exists and is a valid R, and the
requesting user is the instance owner, then set the 'alloc-bypass'
flag, commit R to the KVS, and finally emit the 'alloc' event.