job-manager: add a set-flags event for jobs
Problem: Job manager plugins or other entities may want to be able
to set named "flags" on jobs, e.g. "alloc-bypass" and "debug", but
an operation to change job flags should be recorded as an event in
the eventlog to allow the job state to be properly reconstructed on
eventlog replay.

Add a "set-flags" event with context member `flags` which is an array
of flags to set.