libflux: return routes directly in msg API
Problem: The functions flux_msg_get_route_first() and
flux_msg_get_route_last() return route IDs via an allocated string
that must be freed by the user.  This API made sense previously
when route IDs were stored internally in a zframe.  However, now
that the route IDs are stored in decoded data structures, we could
return the route IDs directly.  In addition, returning this field
as an in-and-out parameter is not consistent to most of the code
in flux-core.

Solution: Return internally stored route IDs via a const char *
return parameter instead of an in-and-out paramter.

Update callers accordingly.