shell: rcalc: use slot size as specified in jobspec
Problem: Jobspec V1 may over-allocate resources to a job if the
job slots are not evenly distributed across a specified number of
nodes. When the shell distributes tasks across a resulting R,
then the shell may segfault in rcalc_distribute(), which derives
the incorrect slot size since it ignores the slot size as specified
in jobspec.

Allow the "cores_per_task" to be specified in rcalc_distribute(),
and use the value specified in the jobspec when calling this function
in the shell. This will ensure that the tasks/slots are properly
distributed even when the scheduler has over-allocated resources.

Fixes #3747