job-manager: allow submit event post in NEW state
Problem: flux job exec service will need to allow the job manager
to post the submit event in place of the job-ingest module, but
events are not permitted to be posted in NEW state to prevent
events from preceding the submit event in the journal.

Allow an exception for the submit event itself.