job-manager: add EVENT_NO_COMMIT flag for use with submit
Problem: The "submit" event posted by the job manager skips use
of event_job_post_pack() because the job-ingest module has already
posted the event to the eventlog. However, this requires the logic
of event_job_post_pack() to be duplicated in submit_post_event(),
which could result in unintended differences in behavior when handling
this event.

Add a EVENT_NO_COMMIT flag which causes event_job_post_pack() to skip
committing the posted event to the eventlog. Update submit_post_event()
to use this flag and remove the duplicated code.