connector/local: implement getopt flux::owner
Problem: it is useful to be able to determine the instance
owner on the other end of a local:// socket without asking
for the security.userid attribute.

Call usock_get_cred () on the server socket once connected,
and cache the result.  Implement the connector getopt() callback
with support for fetching userid portion of the cred via the key
'flux::owner'.