libjob: ask connector for owner ID
Problem: flux_job_submit() avoids signing jobs with munge if
the broker 'security.owner' attribute matches the submitting user.
The attribute cannot be fetched if the broker is not connected to
its upstream peer.  The submit code falls back to munge, but in
the RPM buildfarm environment, the attempt to sign with munge causes
t3302-system-offline.t to fail because the munge daemon is not running.

Ask the connector for the owner ID instead.  The local connector implements
this using SO_PEERCRED, which is not subject to the limitation above.
If the connector cannot supply the owner, fall back to munge.

Fixes #3760