flux-job: refine wait subcommand exit code
Problem: when a job fails because tasks exited with a nonzero
exit code, flux job wait exits with an exit code of 1.

Use flux_job_wait_get_exit_code() and exit with the highest
exit code of the job, 128 + signum if signaled, or 1 if terminated
due to exception.

Add some tests to t2208-job-manager-wait.t to cover the three
scenarios above.