libioencode: encode invalid utf-8 binary data as base64
Problem: If json_pack() encounters invalid utf-8 bytes in ioencode
input, then encoding fails. This can cause unnecessary failures in
jobs that may output binary or other non-UTF-8 data.

If json_pack() of the ioencode data fails, fall back to base64
encoding. Note the encoding in the resulting context object, and
use this in iodecode() to decode base64 data if necessary.

This allows normal data to continue to be utf-8 encoded by default,
and only fall back to the more verbose, but safer, base64 encoding
when there is a problem.