python: fix --require-version processing in jobspec validator
Problem: The jobspec validator plugin does not pass the supplied
--require-version to the flux.job.validate_jobspec() method, nor
does it have a way for the jobspec version to be set to any
version.

Pass the required_version along to validate_jobspec() so its value
is actually used when validating jobspec. Also, allow

 --require-version=any

to disable the version check and allow any canonical jobspec
regardless of version.