Remove excess use of LIBDL
Problem: The LIBDL makefile macro was used excessively.  In many cases
the library was already linked via other libraries or even in other
macros used in the same linker lines.

Solution: Remove LIBDL when it is no longer necessary.  In other cases,
simply add LIBDL to a macro that is already used so it isn't copied
everywhere.