Add missing LIBSODIUM_CFLAGS / LIBS in Makefiles
Problem: In the prior commit we removed excess usage of LIBSODIUM_CFLAGS
and LIBSODIUM_LIBS when code did not directly call code in libsodium.  However,
in the broker, libkvs, and kvs module, LIBSODIUM_CFLAGS and/or LIBSODIUM_LIBS
were not set even though there are calls to libsodium.

Solution: As needed, add LIBSODIUM_CFLAGS or LIBSODIUM_LIBS in the broker,
libkvs, and kvs module.