lua/t0003-events: Mangle payload differently
Problem: In the t0003-events.t file, an invalid base64 payload is
created by removing a padding byte at the end of the payload.  In some
base64 implementations, the lack of a padding byte will not result
in an error.  Therefore the test will fail.

Solution: Mangle the base64 payload by placing an illegal character
in the payload instead of removing the padding byte.