broker: rework fanout options
Problem: there is an immutable tbon.arity attribute and
a -k broker option.  Since many attributes are settable
on the broker command line with -S, it is confusing to have
an attribute that can only be set with an option.  Also,
tbon.arity is a strange name for the TBON fanout.

Drop -k,--k-ary=N broker option.
Rename tbon.arity to tbon.fanout.
Allow tbon.fanout to be set on the command line.

Update users, including tests, the systemd unit file,
and flux-sharness.sh.

Fixes #3795