shell: add plugin for Cray PMI
Interfacing with Cray PMI requires support for libpals,
which means setting a handful of environment variables and
writing out a file containing a set of structs which will
be read in by libpals.

Add a shell plugin for setting the necessary environment variables
and writing out the file.

Fixes flux-framework#3825.