librlist: fix combination of unlike node in rlist_to_R
Problem: When rlist_to_R() compares nodes for possible collapsing,
it checks that each child resource idset in the first node matches
the same child resource idset in the second node. This fails however,
when the 2nd node has extra resource types, since those types will
not be compared. This results in the nodes being considered to
have identical resources, and the extra resources in the 2nd node
are lost.

Ensure that this does not happen by returning nonzero in rnode_cmp()
when the two nodes being compared have a differnt number of child
resource types.

Fixes #3183