broker/overlay: add overlay.topology RPC
Problem: tools may need to know the topology for a subtree,
but tools should not compute the topology assuming the k-ary tree
properties, since irregular trees may be configurable in the future.

Add an RPC 'overlay.topology' which returns a subtree topology object:

{ "rank":N, "size":N, "children":[...] }

Rank is the root of the subtree.
Size is the size of the subtree.
Children is a list of topology objects for each child, if any.

Example for rank 0, size 3, fanout 2:

{
  "rank": 0,
  "size": 3,
  "children": [
    {
      "rank": 1,
      "size": 1,
      "children": []
    },
    {
      "rank": 2,
      "size": 1,
      "children": []
    }
  ]
}