broker: add bootstrap.enable_ipv6 boolean config
Problem: the current default of enabling ipv6 for all zeromq
sockets prevents the flux system instance from binding to an
interface that only has ipv4 addresses.

The enabling of ipv6 mode was merged with PR #3489.  Change the
logic so that the PMI bootstrap code enables ipv6 as before,
but config bootstrap leaves it disabled by default and allows
it to be enabled by configuration:

  [bootstrap]
  enable_ipv6 = true

Update unit test.
Update flux-config-bootstrap(5) man page.

Fixes #3824