broker: avoid global czmq socket options
Problem: the broker sets up some global czmq socket options,
but it is somewhat more clear and flexible if these options
are set explicitly when the sockets are created.

Drop the global socket options and update the places where zeromq
sockets are created by the broker with equivalent socket options.