broker: document 0MQ context situation
Problem: the way that the shmem connector and broker interact
with czmq is a bit subtle if you are not intimate with details of
how libczmq works.

Add some inline comments about the global 0MQ context that libczmq
creates on demand and shares within a process, as well as the atexit
handler for cleaning up unclosed sockets, and libczmq's internal logging.