github: increase ci timeout to 40 minutes
Problem: Over time, the addition of tests to the testsuite has
increased the minimum runtime of the GitHub workflow builds. The
coverage build is especially impacted, presumably due to extra IO
induced by gcov output, and is now often timing out.

Increase the build timeout to 40 minutes to avoid spurious CI timeouts
in the coverage builder.