libflux/fripp: avoid loop on bad FLUX_FRIPP_STATSD
Problem: if FLUX_FRIPP_STATSD is set to an invalid or unknown
host:port, the error is logged repeatedly.

Change the logic so that a fripp context is created even when
the host:port parsing fails, but with an enable flag set to false.
This prevents the stats functions from trying to (re-)create the fripp
context on demand when we know it's going to fail each time.