shell: capture jobspec version in struct jobspec
Problem: The version of the jobspec is not captured in the
struct jobspec (shell->info->jobspec). While the shell currently
only support version 1 jobspec, this may change in the future, so
it will be useful to have the jobspec version number available.

Add a version field to `struct jobspec` and capture the version
key from the jobspec JSON into this member.