t5000-valgrind.t: do not run valgrind test by default
Problem: The t5000-valgrind.t test is susceptible to false positives
when run on new systems and different distros due to the potential
for new false positives from system libraries which have not yet been
added to the suppressions file. This can confuse new users when the
test unnecessarily fails.

Disable the valgrind test by default to avoid these situations.
The test can be run with FLUX_ENABLE_VALGRIND_TEST (e.g. in CI) or
by passing the `-d, --debug` flag (since this is common when running
the test by hand).