broker/overlay: clean up dropped message logging
Problem: code for logging dropped messages is duplicated in
two overlay socket receive callback functions.

Factor out a logdrop() helper function, then change the receive
callbacks to use this function.  This allows a goto label to be
eliminated in both functions, simplifying them slightly.

Note: logging was suppressed for messages received from TBON
children before their hello request.  This should not be possible,
and in fact may indicate that the parent rebooted without notification
getting through to the child.  Log this for now - a future commit
will also send a disconnect message to the child.