flux-overlay: get hostnames from R if missing
Problem: if hostnames are not available in the config.hostlist attribute,
'flux overlay status --hostnames' bails out.

Fall back to fetching resource.R from the KVS if config.hostlist is
not available.  Place this code in a function that is reusable by
multiple subcommands instead of hardwired in the status subcommand.

Add a gethostbyrank subcommand mainly for testing, but possibly
also of use in other contexts.

Update t3303-system-healthcheck.