flux-overlay: add disconnect subcommand
Problem: a misbehaving node may need to be administratively
detached from the flux instance.

Implement 'flux overlay disconnect RANK' which sends a
disconnect-subtree request to RANK's parent, telling it
to lose RANK.

Add 'parentof' subcommand for looking up the parent of a rank
in the overlay topology, mainly for testing, but possibly
useful in other contexts.