docker: add script to build/run system instance testing
Problem: The fluxorama docker image is not ideal for iterative
development.

Add a new script src/test/docker/docker-run-systest.sh, which,
similar to `docker-run-checks.sh -I`, starts a fluxorama docker
container with the current UID/GID registered, mounts the build
directory at /usr/src, and runs an interactive shell in the instance.

The script can be used interactively or in CI to run `make check` or
other tests under the system instance or with systemd present.