broker/overlay: handle upstream broker crash
Problem: if broker crashes without getting word to child,
the child may reconnect and resume messaging once the broker
returns to service.  These messages will be dropped, but
the child is not informed that it should restart.

Send a KEEPALIVE_DISCONNECT in that case, triggering a
subtree panic.

Also, suppress logging dropped messages from children that
have been forceably disconnected, as it is expected that some
messages might be sent before the KEEPALIVE_DISCONNECT is
processed on the other end.

Improve documentation of expected failure scenarios in the
child message handler.

Fixes #3608