libzmqutil: add zmqutil_monitor class
Add a class that wraps zmq_socket_monitor(), to be used for
logging purposes.

The implementation is stubbed for versions <= 4.1.4.
There was a bug in 4.1.4 where a thread race in zmq could deadlock
the flux broker, so earlier use of zmq_socket_monitor() was
removed in pr #3398.

Add a simple unit test for invalid parameters.