testsuite: fix racy disconnect test
Problem: test fails occasionally in CI

This was observed in CI:

  expecting success:
    pid=$(cat health.pid) &&
    echo waiting for pid $pid &&
    test_expect_code 1 wait $pid &&
    grep "overlay disconnect" health.err

  waiting for pid 604451
  not ok 15 - background RPC fails with overlay disconnect (tracker response from 6)

Since the RPC was requested through rank 13, and rank 13 exits
as a result of the crash of broker 6, it may be possible to receive
other errors here, such as ECONNRESET.

Change the test to just expect failure, not a particular one.
The important outcome is that RPCs do not hang.