rpc: deactivate msg_handler when RPC is complete
Problem: RPCs continue to be "active" watchers in a reactor even after
all expected responses have been received. This causes the reactor
to fail to exit (or report > 0 active watchers with FLUX_REACTOR_NOWAIT),
even though the reactor is technically waiting for no more events.

If an RPC is not streaming, stop the msg_handler for the RPC once the
response is received. No more responses are expected, so there is no
reason to keep the msg_handler active anyway.

If there is an error response, stop the msg_handler for both streaming
and non-streaming RPCs, since for FLUX_RPC_STREAMING an error response
terminates the mult-response RPC.

Fixes #3852