docker: add Dockerfile for Fedora Core 34
Problem: Flux is not built and tested on the latest Fedora Core
release with GCC 11.2.

Add a Dockerfile for Fedora 34. It is essentially a copy of the
existing fedora33 Dockerfile. Fedora 34 will be the branch point for
RHEL 9, so this will be a useful platform under which to run CI checks.