libsdprocess: add new systemd launcher library
Problem: In the future we would like to run some processes that can
survive the flux broker crashing.

Solution: As a potential solution we can run those processes under
systemd using the sd-bus library.  We add the libsdprocess library
as a convenience around the sd-bus library for launching and waiting
for process changes using the flux reactor.  A number of convenience
functions have also been added, such as for determining underlying
pid and finding already running processes.