build: shell: create links to shell plugins outside of .libs
Problem: Loading plugins from a subdirectory of src/shell/plugins
is not currently possible when running in-tree, since the default
pluginpath has to point to src/shell/plugins/.libs instead of just
src/shell/plugins/. This is because the plugin loader loads .so
files, which libtool always places in .libs.

Ideally, shell plugins could use the same structure in-tree as when
installed. This would allow optional shell plugins to be installed
in subdirectories of the default shell pluginpath, and loaded
explicitly via

 plugin.load ("foo/bar.so")

For this to work, we'll have to link non builtin shell plugin .so
files into the parent directory of their `.libs`. This is similar
to what is already done for the Lua modules in src/bindings/lua,
so copy that scheme here for use with any eventual shell plugins.