shell: allow plugin.load (pattern) to work as expected
Problem: The shell plugin.load() method requires a table with at
least file=pattern member. However, this is difficult to remember as
well as overly verbose when a configuration object is not required.

If a string is passed as the first argument to plugin.load, treat it
as a plugin to load so that

 plugin.load ("foo.so")

works as expected.