build: prepare for new pmix configure option
Problem: broker builds in a preference for being bootstrapped
with pmix if HAVE_LIBPMIX is defined, but the name is inappropriate
since PMIX may be enabled for the pmix shell plugin, without this
bootstrap preference.

Rename HAVE_PMIX to BROKER_PMIX, and also add conditionals in the
broker Makefile.am to avoid linking with pmix unless BROKER_PMIX is
defined.