jobtap: change prototype of jobtap_stack_call()
Problem: jobtap_stack_call() always calls the main stack of plugins
in jobtap->plugins, but this means the function can't be called on
against an alternate plugin list, which may limit reusability.

Add a new parameter to jobtap_stack_call() -- the list of plugins
on which to invoke the callback. Change all existing callers to
explicitly pass jobtap->plugins for now.