job-manager: add simple job subscription facility
Problem: It may be useful to allow job manager plugins to get more
frequent callbacks on jobs, for example a callback on every job event
instead of just job state changes. However, invoking callbacks for
every jobtap plugin on every job event will be inefficient.

Add a facility to allow plugins to subscribe to one or more jobs to
opt-in to more frequennt callbacks. This change does not actually
enable any more callbacks, but sets up the subscription mechanism.

Since job subscription will be uncommon, and less plugins will
subscribe to a job than are registered with jobtap as a whole, this
should be a much more efficient mechanism for enabling job event
callbacks for plugins.

Becuase either the plugin or the job could be destroyed first, the
plugin keeps a list of subscribed jobs in addition to the list of
plugins subscribed that is added to struct job. This allows a plugin
to remove itself from all subscribed jobs on unload, and a job can
remove itself from all plugins when it becomes inactive.