jobtap: allow plugins to subscribe to all job events
Problem: It may be useful for job manager jobtap plugins to optionally
get callbacks for all job events instead of only job state changes.
For instance, a plugin may want to take action after a `start` event
instead of the transition to RUN state.

Add a new jobtap call

 int flux_jobtap_job_subscribe (flux_plugin_t *p, flux_jobid_t id);

which plugins can optionally use to subscribe to all job events. If one
or more plugins is subscribed to a job, then jobtap will issue
job.event.* callbacks for the job for every job event posted to the
main job eventlog.

A plugin can unsubscribe from a job explicitly with

 void flux_jobtap_job_unsubscribe (flux_plugin_t *p, flux_jobid_t id);

otherwise the subscription will be released when either the plugin or the
job is destroyed.