job-manager: release dependency-after jobs after start event
Problem: after:JOBID dependencies are currently released after the
job enters the RUN state, or after the job "alloc" event. However,
the documentation and probably expectation is that 'after' dependencies
are released after the dependency job has started.

Use job event subscription in the dependency-after plugin to delay
the release of dependent "after" jobs until the "start" event instead
of on the entry to RUN state (after the "alloc" event).

On job submission of an after:JOBID job, ensure that the target job
hasn't already had a "start" event, instead of checking only that the
job is at or past the RUN state.