jobtap: add flux_jobtap_job_event_posted()
Problem: There is no way for a jobtap plugin to quickly determine if
a job has had a given event name posted to its eventlog.

Add flux_jobtap_job_event_posted() to the jobtap interface. This function
will return 1 if an event has been posted to the job eventlog, 0 if not,
or -1 on error.