job-exec: add unload method for exec implementations
Problem: The job-exec module does not issue a final call to all
exec implementations to allow them to destroy any state that was
created during config(). This could lead to memory leaks for future
methods which require creation of such state.

Add an implementation unload() method which is called for all
implementations when the job-exec module is unloaded.