job-exec: add optional override mode for testexec
Problem: The testexec module issues a start event immediately on
reciept of a job start request when mock execution is requested in
the submitted jobspec. However, it may be useful for testing purposes
and otherwise to manually control when the start event is issued.

Add an "override" option to the testexec implementation which can
be enabled via attributes.system.exec.test.override set in jobspec.
When override is enabled, the testexec module does not immediately
issue a start response to the job manager for the test job, but
hashes the job and waits for an external request to issue the
"start" event.

The testexec implementation registers a service endpoint at
job-exec.testexec for control of override jobs, which accepts requests
with payloads of the form

 {
   "jobid":I   #  target jobid
   "event":s   #  event to issue for the target job: "start" or "finish"
   "status"?i  #  for "finish" only, the wait status to report (default: 0)
 }

Note: Override testexec jobs with duration == 0 (unlimited) will not
issue a finish event until a finish request is received or a job exception
is generated for the job in question. Override jobs with either nonzero
duration or attributes.system.exec.test.run_duration will run until the
set duration expires, though a finish RPC may still be issued to stop
the job early.