shell: initrc.lua: support FLUX_SHELL_RC_PATH
Problem: There is no method by which an external package installed
in a different root than flux-core can notify the job shell to load
extra Lua rc scripts and/or plugins at runtime.

Modify the default shell initrc.lua to look for FLUX_SHELL_RC_PATH
in the environment of the job, and if set, treat the value as a colon
separated list of paths to search after the default rcpath (same
directory as initrc.lua) for *.lua rc files to load.

This allows an external package to append its own path to
FLUX_SHELL_RCPATH in order to get a Lua rc script executed and
customize job shell startup. For example, this could be used
to load an external plugin, modify job environment, etc.

Fixes #3866