testsuite: check that after:JOBID jobs are released on start
Problem: dependency=after jobs are now guaranteed not to be released
until the antecedent job's start event, but none of the tests in
t2271-job-dependency-after.t verify that the feature is working.

Add a test to t2271-job-dependency-after.t that ensures dependent
jobs are not released immediately when a job enters the RUN state
(i.e. on an alloc event), but instead release of a job is triggered
when the antecdent receives a start event.