job-manager: add function to globally index event names
Problem: It would be good to capture which events have been posted to
individual jobs, but a per-job structure would require event names
to be duplicated for every job.

Allow for a potentially more memory efficient solution by adding a job
manager global hash of event names to integer index. The solution assigns
a unique, sequential identifier to named events, and allows that id
to be quickly captured.

This will allow an unlimited number of unique event names to be stored
once globally, then referenced by identifier within jobs, greatly
reducing the memory footprint of the final solution.