job-manager: add facility to cache events by id
Problem: Checking if a job has received an event will be useful
for advanced plugin synchronization, but the only way to perform this
check now is by fetching the entire eventlog from the KVS. It would
be much more efficient if each job tracked received events in memory.

Since the event module in the job manager provides a mapping of event
name to unique id, add a bitmap in the job structure to track events
in a very space efficient manner.

The events bitmap has a fixed size of 64 bits, allowing for a total of
64 unique events across a Flux instance. This will almost assuredly be
enough for the time being. The bitmap interface does allow resizing,
however, so if this becomes a problem in the future bitmaps could be
dynamically resized upon seeing the first event with an id of 64 or
greater.