job-manager: track events posted to jobs
Problem: The event and job modules have the framework to track events
posted to jobs, but the tracking is not yet implemented.

Add a function to get an identifier for each event name posted to a job
and store the resulting id in the job event id cache.

Call this function from event_job_post_entry().