shell: add taskids idset to flux_shell_get_rank_info()
Problem: A shell plugin may need to determine the set of task ranks
running on a shell rank, but this information is not exported to plugins.

Add a "taskids" member of the shell "rank info" object returned by
flux_shell_get_rank_info(). This key contains a string of task rank
ids in RFC 22 IDset form for the shell rank that is the target of the
query.

Fixes #3820