doc: update docuemntation for flux_shell_get_rank_info(3)
Problem: A taskids member has been added to the
flux_shell_get_rank_info() payload, but the new field is not
documented.

Add the `taskids` key to the documented contents of the rank info
payload.

Additionally, add a small description of all the fields returned by
this call to increase clarity of the mapage.