testsuite: fix test for flux-python vs PYTHON_INTERPRETER
Problem: The test in t0001-basic.t that ensures flux-python runs
the configured python does not handle the case where the configured
python is a symlink. In this case PYTHON_INTERPRETER and sys.executable
will differ, since Python's sys.executable appears to return the real
path of the Python executable.

Fix the test by comparing sys.executable for both flux-python and
PYTHON_INTERPRETER.